feat: OCR 약 등록을 위한 카메라 촬영 및 통신 로직 구현

+++ app/src/main/java/com/example/bravohealthpark/domain/medication/dto/SendImageResponse.java
@@ -1,12 +1,12 @@
 package com.example.bravohealthpark.domain.medication.dto;
 
-import com.example.bravohealthpark.domain.user.State;
+import com.example.bravohealthpark.domain.medication.State;
 
 public class SendImageResponse {
     private int days;
     private String endDate;
     private String itemName;
-    private int medicationInfoId;
+    private long medicationInfoId;
     private String memo;
     private String startDate;
     private State state;
